fixes & new stuff
fixed: workers/updateServer
new: external services

+++ setup.js
@@ -1,3 +1,12 @@
+////////////////////////////////////////////////
+//     DO YOU WANT TO RESET THE DATABASE?     //
+//        OR SETUP FOR THE FIRST TIME?        //
+////////////////////////////////////////////////
+let reset = true;
+////////////////////////////////////////////////
+
+
+
 const mysql = require('mysql');
 const log = require('leekslazylogger');
 const services = require('./services.json');
@@ -32,23 +41,19 @@ db.connect(function (err) {
 
     log.success(`Connected to database (${database.name}@${database.host})`);
 
-    db.query(`DROP TABLE minecraft;`, (err, result) => {
-        if (err) return log.error(err);
-        // log.debug(result);
-        log.info(`Deleted 'minecraft' table`);
-    });
-
-    // db.query(`DROP TABLE mc_query;`, (err, result) => {
-    //     if (err) return log.error(err);
-    //     // log.debug(result);
-    //     log.info(`Deleted 'mc_query' table`);
-    // });
-    
-    db.query(`DROP TABLE websites;`, (err, result) => {
-        if (err) return log.error(err);
-        // log.debug(result);
-        log.info(`Deleted 'websites' table`);
-    });
+    if (reset) {
+        db.query(`DROP TABLE minecraft;`, (err, result) => {
+            if (err) return log.error(err);
+            // log.debug(result);
+            log.info(`Deleted 'minecraft' table`);
+        });
+        
+        db.query(`DROP TABLE websites;`, (err, result) => {
+            if (err) return log.error(err);
+            // log.debug(result);
+            log.info(`Deleted 'websites' table`);
+        });
+    }
 
     db.query("CREATE TABLE `api`.`minecraft` ( `id` VARCHAR(128) NOT NULL , `name` VARCHAR(128) NOT NULL , `status` VARCHAR(128) NULL , `player_count` SMALLINT NOT NULL , `players` JSON NOT NULL , `tps` DECIMAL(4,2) NOT NULL , `last_online` VARCHAR(32) NULL  , `expires` VARCHAR(32) NULL );", (err, result) => {
         if (err) return log.error(err);
@@ -56,15 +61,7 @@ db.connect(function (err) {
         log.success(`Created 'minecraft' table`);
     });
 
-    // db.query("CREATE TABLE `api`.`mc_query` ( `id` VARCHAR(128) NOT NULL , `name` VARCHAR(128) NOT NULL , `status` VARCHAR(128) NULL , `player_count` SMALLINT NOT NULL , `last_queried` VARCHAR(32) NULL , `last_online` VARCHAR(32) NULL );", (err, result) => {
-    //     if (err) return log.error(err);
-    //     // log.debug(result);
-    //     log.success(`Created 'mc_query' table`);
-    // });
-
-    
-
-    db.query("CREATE TABLE `api`.`websites` ( `id` VARCHAR(128) NOT NULL , `name` VARCHAR(128) NOT NULL , `status` VARCHAR(128) NULL , `last_queried` VARCHAR(32) NULL , `last_online` VARCHAR(32) NULL );", (err, result) => {
+    db.query("CREATE TABLE `api`.`websites` ( `id` VARCHAR(128) NOT NULL , `name` VARCHAR(128) NOT NULL , `status` VARCHAR(128) NULL , `last_online` VARCHAR(32) NULL );", (err, result) => {
         if (err) return log.error(err);
         // log.debug(result);
         log.success(`Created 'websites' table`);
@@ -94,15 +91,19 @@ db.connect(function (err) {
         }); 
     };
 
-    // db.query(`INSERT INTO mc_query (id, name, player_count) VALUES ('proxy', '${services.minecraft.servers.proxy.name}', 0);`, (err, result) => {
-    //     if (err) return log.error(err);
-    //     // log.debug(result);
-    //     log.console(`Added '${services.minecraft.servers.proxy.name}' to the 'mc_query' table`);
-    //     log.info("Done")
-    // });
+    // add external
+    let external = services.external;
+    for (api in external) {
+        let a = api;
+        db.query(`INSERT INTO websites (id, name) VALUES ('${a}', '${external[a].name}');`, (err, result) => {
+            if (err) return log.error(err);
+            // log.debug(result);
+            log.console(`Added '${external[a].name}' external API to the 'websites' table`);
+        }); 
+    };
 
 
-    log.info("Setup process does not automatically exit after it is finished (Ctrl+C)", 'yellowBright')
+    log.info("Setup process does not automatically exit after it is finished (Ctrl+C)", 'yellowBright');
 
 });
